+++ src/leanify/zip.h
@@ -7,7 +7,7 @@ class Zip {
  public:
   explicit Zip(void* p, size_t s) : fp_(static_cast<uint8_t*>(p)), size_(s) {}
 
-  size_t Leanify(const ECTOptions& Options, unsigned long* files);
+  size_t Leanify(const ECTOptions& Options, size_t* files);
   uint32_t RecompressFile(unsigned char* data, uint32_t size, uint32_t size_leanified, std::string filename, const ECTOptions& Options);
 
   static const uint8_t header_magic[4];
